Revert "change extern variable type"
This reverts commit 67f902589a0a90d84a5d1ec8a4f5db6bb56574ac.

+++ main.c
@@ -1,6 +1,6 @@
  #include "monty.h"
 
-int glob[] = {0, 0, 0};
+long int glob[] = {0, 0, 0};
 /**
  * main - the main execution frame for monty interpreter
  * @argc: number of arguments read from command line
